requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.2.2
idna==2.7
noise==1.2.2
numpy==1.14.5
Pillow==5.1.0
protobuf==3.5.2.post1
pycparser==2.18
pygame==1.9.3
PyNDN==2.8b1
six==1.11.0
sortedcontainers==2.0.4
appdirs==1.4.3
